fix: resolve nilerr in engine_reader.go
